Tools/requirements.txt
about-time==3.1.1
alive-progress==2.4.1
colorama==0.4.6
grapheme==0.6.0
pyserial==3.5
